requirements.txt
beautifulsoup4==4.10.0
biopython==1.79
blis==0.7.4
catalogue==2.0.6
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
crossref-commons==0.0.7
cymem==2.0.5
filelock==3.2.0
huggingface-hub==0.0.17
idna==3.2
importlib-metadata==4.8.1
Jinja2==3.0.1
joblib==1.0.1
Levenshtein==0.15.0
MarkupSafe==2.0.1
murmurhash==1.0.5
nltk==3.6.4
numpy==1.21.2
packaging==21.0
pandas==1.3.3
pathy==0.6.0
Pillow==8.3.2
preshed==3.0.5
pydantic==1.8.2
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
rapidfuzz==1.6.2
ratelimit==2.2.1
regex==2021.9.30
requests==2.26.0
sacremoses==0.0.46
scikit-learn==1.0
scipy==1.7.1
sentence-transformers==2.1.0
sentencepiece==0.1.96
six==1.16.0
smart-open==5.2.1
spacy==3.1.3
spacy-legacy==3.0.8
srsly==2.4.1
thinc==8.0.10
threadpoolctl==2.2.0
tokenizers==0.10.3
torch==1.9.1
torchvision==0.10.1
tqdm==4.62.3
transformers==4.11.2
typer==0.4.0
typing-extensions==3.10.0.2
urllib3==1.26.7
wasabi==0.8.2
zipp==3.6.0
